refactor: 비커 게임 로직을 useBeakerGame 훅으로 분리
